backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
psycopg[binary]>=3.1.0
